Add 'Folder relation' input to the 'Search files'
