+++ Apps.AmazonS3/Models/Request/SearchFilesRequest.cs
@@ -1,5 +1,7 @@
 ﻿using Apps.AmazonS3.DataSourceHandlers;
+using Apps.AmazonS3.DataSourceHandlers.Static;
 using Blackbird.Applications.Sdk.Common;
+using Blackbird.Applications.Sdk.Common.Dictionaries;
 using Blackbird.Applications.Sdk.Common.Dynamic;
 
 namespace Apps.AmazonS3.Models.Request;
@@ -8,5 +10,9 @@ public class SearchFilesRequest
 {
     [Display("Folder ID (prefix)", Description = "Describe the folder path structure, f.e. /example/")]
     [DataSource(typeof(FolderDataHandler))]
-    public string? FolderID { get; set; }
+    public string? FolderId { get; set; }
+
+    [Display("Folder relation")]
+    [StaticDataSource(typeof(FolderRelationTriggerDataHandler))]
+    public string? FolderRelationTrigger { get; set; }
 }
